requirements.txt
absl-py==0.10.0
argon2-cffi==21.1.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.1.1
certifi==2021.10.8
cffi==1.15.0
cloudpickle==2.0.0
cvxopt==1.2.7
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
ecos==2.0.7.post1
entrypoints==0.3
fonttools==4.28.2
google-auth==1.21.3
google-auth-oauthlib==0.4.1
grpcio==1.32.0
gym==0.21.0
importlib-resources==5.4.0
ipykernel==6.5.1
ipython==7.30.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.6
numpy==1.21.4
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
PICOS==2.3.1
Pillow==8.4.0
pip==21.2.4
prometheus-client==0.12.0
prompt-toolkit==3.0.23
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.0
Pygments==2.10.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
pytorch-lightning==0.9.0
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.2.1
QtPy==1.11.2
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==58.0.4
setuptools-scm==6.3.2
six==1.16.0
tensorboard==2.2.0
tensorboard-plugin-wit==1.7.0
terminado==0.12.1
testpath==0.5.0
tomli==1.2.2
tornado==6.1
tqdm==4.49.0
traitlets==5.1.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.37.0
widgetsnbextension==3.5.2
zipp==3.6.0
